+++ packages/webapp/src/config/AppConfig.ts
@@ -60,7 +60,7 @@ export interface PluginManager {
  * Customize pages
  */
 export interface Pages {
-  disabled?: (("edit" | "date" | "video" | "map" | "tag" | "album" | "all") | string)[];
+  disabled?: (("edit" | "date" | "video" | "map" | "tag" | "album" | "onThisDay" | "all") | string)[];
   /**
    * Initial page of the webapp on its first load. It falls back to the page of all media if the configured page is disabled
    */